+++ apps/client/src/app/core/alarms/+state/alarms.actions.ts
@@ -1,3 +1,4 @@
+import { UpdateData } from '@angular/fire/firestore';
 import { Action } from '@ngrx/store';
 import { Alarm } from '../alarm';
 import { AlarmGroup } from '../alarm-group';
@@ -10,6 +11,8 @@ export enum AlarmsActionTypes {
   AddAlarmsAndGroup = '[Alarms] Add alarms and group',
   AlarmsCreated = '[Alarms] Alarms created',
   UpdateAlarm = '[Alarms] Update Alarm',
+  PureUpdateAlarm = '[Alarms] Pure Update Alarm',
+  SetAlarmDone = '[Alarms] Set Alarm Done',
   RemoveAlarm = '[Alarms] Remove Alarm',
 
   PersistAlarms = '[Alarms] Persist Alarms',
@@ -63,6 +66,20 @@ export class UpdateAlarm implements Action {
   readonly type = AlarmsActionTypes.UpdateAlarm;
 
   constructor(public readonly alarm: Alarm) {
+  }
+}
+
+export class PureUpdateAlarm implements Action {
+  readonly type = AlarmsActionTypes.PureUpdateAlarm;
+
+  constructor(public readonly key: string, public readonly payload: UpdateData<Alarm>) {
+  }
+}
+
+export class SetAlarmDone implements Action {
+  readonly type = AlarmsActionTypes.SetAlarmDone;
+
+  constructor(public readonly key: string, public readonly done: boolean) {
   }
 }
 
@@ -153,4 +170,6 @@ export type AlarmsAction =
   | LoadAlarmGroup
   | AlarmGroupLoaded
   | SetAlarms
-  | SetGroups;
+  | SetGroups
+  | PureUpdateAlarm
+  | SetAlarmDone;