+++ apps/client/src/app/core/alarms/+state/alarms.effects.ts
@@ -11,7 +11,9 @@ import {
   CreateAlarmGroup,
   DeleteAlarmGroup,
   LoadAlarmGroup,
+  PureUpdateAlarm,
   RemoveAlarm,
+  SetAlarmDone,
   SetAlarms,
   SetGroups,
   UpdateAlarm,
@@ -59,6 +61,14 @@ export class AlarmsEffects {
     withLatestFrom(this.alarmsFacade.allAlarms$),
     switchMap(([, alarms]) => {
       return this.alarmsService.deleteAll(alarms);
+    })
+  ), { dispatch: false });
+
+
+  markAlarmAsDone$ = createEffect(() => this.actions$.pipe(
+    ofType<SetAlarmDone>(AlarmsActionTypes.SetAlarmDone),
+    map(({ key, done }) => {
+      return this.alarmsService.setAlarmDone(key, done);
     })
   ), { dispatch: false });
 
@@ -164,6 +174,13 @@ export class AlarmsEffects {
     .pipe(
       ofType(AlarmsActionTypes.UpdateAlarm),
       switchMap((action: UpdateAlarm) => this.alarmsService.update(action.alarm.$key, action.alarm))
+    ), { dispatch: false });
+
+
+  pureUpdateAlarmInDatabase$ = createEffect(() => this.actions$
+    .pipe(
+      ofType(AlarmsActionTypes.PureUpdateAlarm),
+      switchMap((action: PureUpdateAlarm) => this.alarmsService.pureUpdate(action.key, action.payload))
     ), { dispatch: false });