+++ apps/client/src/app/core/alarms/+state/alarms.facade.ts
@@ -13,7 +13,9 @@ import {
   DeleteAllAlarms,
   LoadAlarmGroup,
   LoadAlarms,
+  PureUpdateAlarm,
   RemoveAlarm,
+  SetAlarmDone,
   SetAlarms,
   SetGroups,
   UpdateAlarm,
@@ -43,6 +45,7 @@ import { safeCombineLatest } from '../../rxjs/safe-combine-latest';
 import { LazyDataFacade } from '../../../lazy-data/+state/lazy-data.facade';
 import { LazyData } from '../../../lazy-data/lazy-data';
 import { XivapiPatch } from '../../data/model/xivapi-patch';
+import { UpdateData } from '@angular/fire/firestore';
 
 @Injectable({
   providedIn: 'root'
@@ -462,8 +465,14 @@ export class AlarmsFacade {
 
   private sortAlarmDisplays(alarms: AlarmDisplay[]): AlarmDisplay[] {
     return alarms.sort((a, b) => {
+      if (a.spawned && a.alarm.done) {
+        return 1;
+      }
+      if (b.spawned && b.alarm.done) {
+        return -1;
+      }
       if (a.spawned && b.spawned) {
-        return a.remainingTime < b.remainingTime ? -1 : 1;
+        return Math.abs(a.remainingTime - b.remainingTime) < 10 ? -1 : 1;
       }
       if (a.spawned) {
         return -1;
@@ -471,7 +480,7 @@ export class AlarmsFacade {
       if (b.spawned) {
         return 1;
       }
-      if (a.remainingTime === b.remainingTime) {
+      if (Math.abs(a.remainingTime - b.remainingTime) < 10) {
         return a.alarm.itemId < b.alarm.itemId ? -1 : 1;
       }
       return a.remainingTime < b.remainingTime ? -1 : 1;
@@ -592,5 +601,13 @@ export class AlarmsFacade {
       return [alarm, { ...alarmWithFishEyesEnabled, fishEyes: true }];
     }
     return [alarm];
+  }
+
+  pureUpdateAlarm(key: string, alarm: UpdateData<Alarm>): void {
+    this.store.dispatch(new PureUpdateAlarm(key, alarm));
+  }
+
+  setAlarmDone(key: string, done: boolean): void {
+    this.store.dispatch(new SetAlarmDone(key, done));
   }
 }