+++ apps/client/src/app/core/electron/ipc.service.ts
@@ -15,7 +15,8 @@ import { FreeCompanyDialog, Message } from '@ffxiv-teamcraft/pcap-ffxiv';
 import { toIpcData } from '../rxjs/to-ipc-data';
 import { UpdateInstallPopupComponent } from '../../modules/ipc-popups/update-install-popup/update-install-popup.component';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
-import { NzMessageRef, NzMessageService } from 'ng-zorro-antd/message';
+import { NzMessageService } from 'ng-zorro-antd/message';
+import { PacketCaptureStatus } from './packet-capture-status';
 
 type EventCallback = (event: IpcRendererEvent, ...args: any[]) => void;
 
@@ -44,6 +45,8 @@ declare global {
 export class IpcService {
 
   public static readonly ROTATION_DEFAULT_DIMENSIONS = { x: 600, y: 200 };
+
+  public readonly pcapStatus$ = new BehaviorSubject<PacketCaptureStatus>(PacketCaptureStatus.STOPPED);
 
   public packets$ = new Subject<Message>();
 
@@ -380,7 +383,7 @@ export class IpcService {
       });
     });
     this.on('metrics:imported', () => {
-      this.message.remove()
+      this.message.remove();
       this.message.info(this.translate.instant('METRICS.Imported'), {
         nzDuration: 10000
       });
@@ -470,14 +473,28 @@ export class IpcService {
           }
         });
     });
+    this.on('pcap:status', (e, status) => this.pcapStatus$.next(status));
     // If we don't get a packet for an entire minute, something is wrong.
     this.packets$.pipe(
       debounceTime(60000)
     ).subscribe(() => {
+      this.pcapStatus$.next(PacketCaptureStatus.WARNING);
       this.send('log', {
         level: 'error',
         data: 'No ping received from the server during 60 seconds'
       });
+    });
+    // If we don't get a packet for 5 minutes, attempt to restart pcap entirely.
+    this.packets$.pipe(
+      debounceTime(300000)
+    ).subscribe(() => {
+      this.pcapStatus$.next(PacketCaptureStatus.WARNING);
+      this.send('log', {
+        level: 'error',
+        data: 'No ping received from the server during 300 seconds, restarting pcap'
+      });
+      this.send('toggle-machina', false);
+      this.send('toggle-machina', true);
     });
     this.handleOverlayChange();
   }
@@ -512,6 +529,9 @@ export class IpcService {
   private handleMessage(packet: Message): void {
     // If we're inside an overlay, don't do anything with the packet, we don't care.
     this.totalPacketsHandled++;
+    if (this.pcapStatus$.value !== PacketCaptureStatus.RUNNING) {
+      this.pcapStatus$.next(PacketCaptureStatus.RUNNING);
+    }
     this.packets$.next(packet);
     const debugPackets = (<any>window).debugPackets;
     if (debugPackets === true || (typeof debugPackets === 'function' && debugPackets(packet))) {