+++ apps/client/src/app/modules/list/+state/lists.reducer.ts
@@ -2,7 +2,7 @@ import { ListsAction, ListsActionTypes } from './lists.actions';
 import { List } from '../model/list';
 import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
 import { ListController } from '../list-controller';
-import { cloneDeep } from 'lodash';
+import structuredClone from '@ungap/structured-clone';
 
 
 const PINNED_LIST_LS_KEY = 'lists:pinned';
@@ -42,7 +42,7 @@ function updateLists(lists: List[], state: ListsState, matchingPredicate = (list
   const afterMap = listsAdapter.map(storeList => {
     checkedLists[storeList.$key] = true;
     const patch = listsByKey[storeList.$key];
-    if (patch && patch.etag > storeList.etag) {
+    if (patch && (patch.etag > storeList.etag || storeList.offline)) {
       if (storeList.$key === state.selectedId) {
         return storeList;
       }
@@ -97,12 +97,11 @@ export function listsReducer(
       const listId = action.listId || state.selectedId;
       const list = ListController.clone(state.listDetails.entities[listId], true);
       const item = ListController.getItemById(list, action.itemId, !action.finalItem, action.finalItem);
-      const requiredHq = ListController.requiredAsHQ(list, item) > 0;
       let fill = true;
-      if (state.autocompletionEnabled && action.settings.enableAutofillHQFilter && requiredHq) {
+      if (state.autocompletionEnabled && action.settings.enableAutofillHQFilter && item.requiredHQ) {
         fill = !action.fromPacket || action.hq;
       }
-      if (state.autocompletionEnabled && action.settings.enableAutofillNQFilter && !requiredHq) {
+      if (state.autocompletionEnabled && action.settings.enableAutofillNQFilter && !item.requiredHQ) {
         fill = !action.fromPacket || !action.hq;
       }
       if (fill) {
@@ -182,14 +181,13 @@ export function listsReducer(
     }
 
     case ListsActionTypes.ListDetailsLoaded: {
-      const newVersion = ListController.clone(cloneDeep(action.payload) as List, true);
-      let updated = false;
+      const newVersion = ListController.clone(structuredClone(action.payload) as List, true);
       let listDetails = state.listDetails;
-      if (!action.forOverlay && state.listDetails.entities[action.payload.$key]) {
+      if (state.listDetails.entities[action.payload.$key]) {
         listDetails = listsAdapter.mapOne({
           id: newVersion.$key,
           map: current => {
-            updated = (newVersion.etag || 0) > (current.etag || 0);
+            const updated = action.forOverlay || (newVersion.etag || 0) > (current.etag || 0) || current.offline;
             if (updated) {
               if (newVersion.items?.length > 0 && newVersion.notFound) {
                 newVersion.notFound = false;
@@ -201,7 +199,6 @@ export function listsReducer(
         }, state.listDetails);
       } else {
         listDetails = listsAdapter.setOne(newVersion, state.listDetails);
-        updated = true;
       }
       state = {
         ...state,