+++ apps/client/src/environments/patch-notes.ts
@@ -1,14 +1,18 @@
 export const patchNotes = `### Bug Fixes
 
-* **gathering-location:** fixed search not finding anything with "Any" node type filter.
-* **metrics:** fixed histogram component not loading.
+* **core:** add structuredClone polyfill for older browsers.
+* **db:** fixed some content not being accessible due to italic characters in their name.
+* **desktop:** fixed omicron tokens listed in island inventory.
+* **food-picker:** now sorting results by picked stat(s) value.
+* **lists:** possible fix for final items sometimes not updating.
+* **step-by-step:** added masterbook icon next to input when required.
+* **step-by-step:** fixed arrows not accurate at all on ultrawide displays.
+* **teams:** fixed discord webhook often sending wrong values in notification.
+* **workshop:** improvements on list deletion workshop update.
 
 
 ### Features
 
-* **db:** allagan reports improt.
-
-
-### Performance Improvements
-
-* **lists:** first step of update system refactoring.`;
+* **db:** import of ~500 allagan reports.
+* **gearset:** new button to reset materia and item progression.
+* **gearset:** you can now switch from a class to its job on a gearset.`;